+++ src/main/java/hr/algebra/webshop/controller/AdminMerchController.java
@@ -8,8 +8,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
 import java.io.IOException;
-import java.util.*;
+import java.util.List;
+import java.util.Objects;
 
 import static hr.algebra.webshop.controller.AuthenticationController.authenticatedShopUser;
 import static hr.algebra.webshop.utils.ImageUtils.imageConvertToBase64ForMerch;
@@ -28,7 +30,7 @@ public class AdminMerchController {
     @GetMapping("/adminMerch")
     public String getAdminMerch(Model model) {
         model.addAttribute("AuthenticatedShopUser", authenticatedShopUser);
-        if (authenticatedShopUser.isAuthenticated()) {
+        if (authenticatedShopUser.isAuthenticated() && authenticatedShopUser.getUserRoleId() == 1) {
             model.addAttribute("merchList", merchService.getAllMerchItems());
             return "adminMerch";
         }
@@ -42,8 +44,11 @@ public class AdminMerchController {
     @GetMapping("/adminMerch/createNewMerch")
     public String createNewMerch(Model model) {
         model.addAttribute("AuthenticatedShopUser", authenticatedShopUser);
-        model.addAttribute("newMerch", new Merch());
-        return "createNewMerch";
+        if (authenticatedShopUser.isAuthenticated() && authenticatedShopUser.getUserRoleId() == 1) {
+            model.addAttribute("newMerch", new Merch());
+            return "createNewMerch";
+        }
+        return "redirect:/dragonBallStore";
     }
 
     @PostMapping("/adminMerch/createNewMerch")
@@ -63,23 +68,29 @@ public class AdminMerchController {
     @GetMapping("/adminMerch/deleteMerch/{idMerch}")
     public String deleteMerch(@PathVariable Long idMerch, Model model) {
         model.addAttribute("AuthenticatedShopUser", authenticatedShopUser);
-        if (idMerch == null) {
+        if (authenticatedShopUser.isAuthenticated() && authenticatedShopUser.getUserRoleId() == 1) {
+            if (idMerch == null) {
+                return "redirect:/adminMerch";
+            }
+            merchService.deleteMerchById(idMerch);
             return "redirect:/adminMerch";
         }
-        merchService.deleteMerchById(idMerch);
-        return "redirect:/adminMerch";
+        return "redirect:/dragonBallStore";
     }
 
     @GetMapping("/adminMerch/editMerch/{idMerch}")
     public String editMerch(@PathVariable Long idMerch, Model model) {
         model.addAttribute("AuthenticatedShopUser", authenticatedShopUser);
-        Merch merch = merchService.getSingleMerch(idMerch);
-        if (merch == null) {
-            model.addAttribute("Product", "Product with id " + idMerch + " not found!");
-            return "notFound";
+        if (authenticatedShopUser.isAuthenticated() && authenticatedShopUser.getUserRoleId() == 1) {
+            Merch merch = merchService.getSingleMerch(idMerch);
+            if (merch == null) {
+                model.addAttribute("Product", "Product with id " + idMerch + " not found!");
+                return "notFound";
+            }
+            model.addAttribute("editMerch", merch);
+            return "editMerch";
         }
-        model.addAttribute("editMerch", merch);
-        return "editMerch";
+        return "redirect:/dragonBallStore";
     }
 
     @PostMapping("/adminMerch/editMerch/{idMerch}")
